moved DummyRequest into the test suite
