+++ piped/processors/test/test_web_processors.py
@@ -12,7 +12,7 @@ from twisted.web import http
 
 from piped import exceptions
 from piped.processors import web_processors
-from piped.providers import web_provider
+from piped.providers.test import test_web_provider
 
 
 class StubRequest(object):
@@ -193,7 +193,7 @@ class TestExtractRequestArguments(unittest.TestCase):
 
     @defer.inlineCallbacks
     def test_simple_extract(self):
-        request = web_provider.DummyRequest([])
+        request = test_web_provider.DummyRequest([])
         request.addArg('foo', 42)
         request.args['bar'] = [1, 2] # a multivalued argument
 
@@ -206,7 +206,7 @@ class TestExtractRequestArguments(unittest.TestCase):
 
     @defer.inlineCallbacks
     def test_get_all_arguments(self):
-        request = web_provider.DummyRequest([])
+        request = test_web_provider.DummyRequest([])
         request.addArg('foo', 42)
         request.args['bar'] = [1, 2] # a multivalued argument
 
@@ -219,7 +219,7 @@ class TestExtractRequestArguments(unittest.TestCase):
 
     @defer.inlineCallbacks
     def test_nonexistent(self):
-        request = web_provider.DummyRequest([])
+        request = test_web_provider.DummyRequest([])
         request.addArg('foo', 42)
 
         baton = dict(request=request)
@@ -231,7 +231,7 @@ class TestExtractRequestArguments(unittest.TestCase):
 
     @defer.inlineCallbacks
     def test_nonexistent_without_skipping(self):
-        request = web_provider.DummyRequest([])
+        request = test_web_provider.DummyRequest([])
         request.addArg('foo', 42)
 
         baton = dict(request=request)
@@ -243,7 +243,7 @@ class TestExtractRequestArguments(unittest.TestCase):
 
     @defer.inlineCallbacks
     def test_nonexistent_with_fallback(self):
-        request = web_provider.DummyRequest([])
+        request = test_web_provider.DummyRequest([])
         request.addArg('foo', 42)
 
         baton = dict(request=request)
@@ -255,7 +255,7 @@ class TestExtractRequestArguments(unittest.TestCase):
 
     @defer.inlineCallbacks
     def test_loading_json(self):
-        request = web_provider.DummyRequest([])
+        request = test_web_provider.DummyRequest([])
         request.addArg('foo', 42)
         request.args['bar'] = ['{"loaded": 42}', '{"loaded": 93}'] # a multivalued argument containing json
 
@@ -267,7 +267,7 @@ class TestExtractRequestArguments(unittest.TestCase):
 
     @defer.inlineCallbacks
     def test_loading_invalid_json(self):
-        request = web_provider.DummyRequest([])
+        request = test_web_provider.DummyRequest([])
         request.addArg('foo', '')
 
         baton = dict(request=request)
